#  x64 backend: add lowerings with load-op-store fusion.