--- a/cranelift/codegen/src/isa/x64/lower/isle/generated_code.manifest
+++ b/cranelift/codegen/src/isa/x64/lower/isle/generated_code.manifest
@@ -1,4 +1,4 @@
 src/clif.isle 443b34b797fc8ace
 src/prelude.isle d8a93eb727abd7f4
-src/isa/x64/inst.isle 79c7275fd5b46560
-src/isa/x64/lower.isle b7fe1c95c21edbe4
+src/isa/x64/inst.isle 6dcba190988a695
+src/isa/x64/lower.isle b95161bdf07b9365