These lowerings use the `OP [mem], reg` forms (or in AT&T syntax, `OP %reg, (mem)`) --
i.e., x86 instructions that load from memory, perform
an ALU operation, and store the result, all in one instruction. Using
these instruction forms, we can merge three CLIF ops together: a load,
an arithmetic operation, and a store.

Stacks on top of #4069; last two commits are new.