@cfallin 2022-04-26
I plan to measure perf impacts using Sightglass tomorrow.