@cfallin 2022-04-26
> Out of curiosity, have you verified that this properly fuses our `externref` write barriers?

I just took a look, and it seems that this (it looks like the inc/dec is [here](https://github.com/bytecodealliance/wasmtime/blob/164bfeaf7e0bbe60c63e3c605266c8bc5edcfd60/crates/cranelift/src/func_environ.rs#L316)?) is actually an atomic operation, so this PR does not handle it. There are also direct-on-memory forms for atomic add/sub (`lock add [mem], reg/imm`) which we could do if this becomes important for performance.